qlog::writer

Function make_qlog_writer_from_path

pub fn make_qlog_writer_from_path<P: AsRef<Path>>(
    path: P,
    compression: QlogCompression,
) -> Result<QlogFileWriter>

Convenience function to create a File at path with a writer based on compression.

Equivalent to manually creating a File and passing it to make_qlog_writer.